NOJIRA: Confirm paths exist before deleting during blueprint cleanup (#601)
* fix: confirm paths exist before deleting

* docs: add changelog entry

+++ includes/blueprints/import.php
@@ -561,6 +561,11 @@ function import_options( array $options ): void {
  * @return void
  */
 function cleanup( $blueprint_zip_path, $blueprint_folder_path ) {
-	wp_delete_file( $blueprint_zip_path );
-	wp_delete_file( $blueprint_folder_path . '/acm.json' );
+	if ( file_exists( $blueprint_zip_path ) ) {
+		wp_delete_file( $blueprint_zip_path );
+	}
+
+	if ( file_exists( $blueprint_folder_path . '/acm.json' ) ) {
+		wp_delete_file( $blueprint_folder_path . '/acm.json' );
+	}
 }
